Refatorando código para Controller do padrão MVC
