+++ 05-express-mvc/routes/admin.routes.js
@@ -1,28 +1,10 @@
 const { Router } = require('express');
-const path = require('path');
+
+const { getAddProduct, postAddProduct } = require('../controllers/productController');
 const router = Router();
-const rootDir = require('../helpers/path');
 
-const products = [];
+router.get('/add-product', getAddProduct);
 
-router.get('/add-product', (req, res, next) => {
-  console.log('💰');
-  res.render('add-product', {
-    pageTitle: 'Novos produtos',
-    path: '/admin/add-product',
-    activeProduct: true,
-    productCSS: true,
-    formsCSS: true,
-  });
-});
+router.post('/add-product', postAddProduct);
 
-router.post('/add-product', (req, res) => {
-  console.log('💰', req.body);
-  products.push({ title: req.body.title });
-  res.redirect('/');
-});
-
-module.exports = {
-  router,
-  products,
-};
+module.exports = router;
